requirements.txt
python-telegram-bot==20.3
scikit-learn==1.3.0
nltk==3.8.1
numpy==1.24.3
SpeechRecognition==3.10.0
gTTS==2.3.2
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
natasha==1.5.0
python-Levenshtein==0.21.1
pydub==0.25.1
ffmpeg-python==0.2.0